requirements.txt

tensorflow>=2.14.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1
scikit-image>=0.22.0
